Surface some Conditions in Block Visibility (#1084) (for phasing out display contexts)
* let 'node_has_term' and 'node_is_islandora_object' be in block visibility conditions

* add config field to NodeIsIslandoraObject condition

Conditions only get applied in Block Visibility when there's a config
change for that condition in the block config form, so it needs a field
to configure.

* more informative field description

* some cleanup to config for NodeIsIslandoraObject

Much thanks to @adam-vessey for helping to figure this out!

* cleanup default config for MediaSourceHasMimetype

Might as well get this cleaned up so it stops inserting itself in all
the block configs where it's not relevant.

* remove unused use statement